Fix stream file health check in Kubernetes
