+++ hedera-mirror-importer/src/main/java/com/hedera/mirror/importer/config/MirrorImporterConfiguration.java
@@ -20,12 +20,9 @@ package com.hedera.mirror.importer.config;
  * ‍
  */
 
-import java.net.URI;
-import java.time.Duration;
 import javax.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
@@ -39,20 +36,10 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.retry.annotation.EnableRetry;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
-import software.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import software.amazon.awssdk.core.interceptor.Context;
-import software.amazon.awssdk.core.interceptor.ExecutionAttributes;
-import software.amazon.awssdk.core.interceptor.ExecutionInterceptor;
-import software.amazon.awssdk.http.SdkHttpRequest;
-import software.amazon.awssdk.http.async.SdkAsyncHttpClient;
-import software.amazon.awssdk.http.nio.netty.NettyNioAsyncHttpClient;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3AsyncClient;
-import software.amazon.awssdk.services.s3.S3AsyncClientBuilder;
 
 import com.hedera.mirror.importer.MirrorProperties;
-import com.hedera.mirror.importer.downloader.CommonDownloaderProperties;
 import com.hedera.mirror.importer.leader.LeaderAspect;
+import com.hedera.mirror.importer.leader.LeaderService;
 
 @Configuration
 @EnableAsync
@@ -62,9 +49,6 @@ import com.hedera.mirror.importer.leader.LeaderAspect;
 public class MirrorImporterConfiguration {
 
     private final MirrorProperties mirrorProperties;
-    private final CommonDownloaderProperties downloaderProperties;
-    private final MetricsExecutionInterceptor metricsExecutionInterceptor;
-    private final AwsCredentialsProvider awsCredentialsProvider;
 
     @Autowired(required = false)
     @Qualifier("webServerStartStop")
@@ -80,58 +64,14 @@ public class MirrorImporterConfiguration {
 
     @Bean
     @Profile("kubernetes")
-    LeaderAspect leaderAspect() {
+    LeaderService leaderAspect() {
         return new LeaderAspect();
     }
 
     @Bean
-    @ConditionalOnProperty(prefix = "hedera.mirror.importer.downloader", name = "cloudProvider", havingValue = "GCP")
-    public S3AsyncClient gcpCloudStorageClient() {
-        log.info("Configured to download from GCP with bucket name '{}'", downloaderProperties.getBucketName());
-        // Any valid region for aws client. Ignored by GCP.
-        S3AsyncClientBuilder clientBuilder = asyncClientBuilder("us-east-1")
-                .endpointOverride(URI.create(downloaderProperties.getCloudProvider().getEndpoint()));
-        String projectId = downloaderProperties.getGcpProjectId();
-        if (StringUtils.isNotBlank(projectId)) {
-            clientBuilder.overrideConfiguration(builder -> builder.addExecutionInterceptor(new ExecutionInterceptor() {
-                @Override
-                public SdkHttpRequest modifyHttpRequest(
-                        Context.ModifyHttpRequest context, ExecutionAttributes executionAttributes) {
-                    return context.httpRequest().toBuilder()
-                            .appendRawQueryParameter("userProject", projectId).build();
-                }
-            }));
-        }
-        return clientBuilder.build();
-    }
-
-    @Bean
-    @ConditionalOnProperty(prefix = "hedera.mirror.importer.downloader", name = "cloudProvider", havingValue = "S3",
-            matchIfMissing = true)
-    public S3AsyncClient s3CloudStorageClient() {
-        log.info("Configured to download from S3 in region {} with bucket name '{}'",
-                downloaderProperties.getRegion(), downloaderProperties.getBucketName());
-        S3AsyncClientBuilder clientBuilder = asyncClientBuilder(
-                downloaderProperties.getRegion());
-        String endpointOverride = downloaderProperties.getEndpointOverride();
-        if (endpointOverride != null) {
-            log.info("Overriding s3 client endpoint to {}", endpointOverride);
-            clientBuilder.endpointOverride(URI.create(endpointOverride));
-        }
-        return clientBuilder.build();
-    }
-
-    private S3AsyncClientBuilder asyncClientBuilder(String region) {
-        SdkAsyncHttpClient httpClient = NettyNioAsyncHttpClient.builder()
-                .maxConcurrency(downloaderProperties.getMaxConcurrency())
-                .connectionMaxIdleTime(Duration.ofSeconds(5))  // https://github.com/aws/aws-sdk-java-v2/issues/1122
-                .build();
-
-        return S3AsyncClient.builder()
-                .credentialsProvider(awsCredentialsProvider)
-                .region(Region.of(region))
-                .httpClient(httpClient)
-                .overrideConfiguration(c -> c.addExecutionInterceptor(metricsExecutionInterceptor));
+    @Profile("!kubernetes")
+    LeaderService leaderService() {
+        return Boolean.TRUE::booleanValue; // Leader election not available outside Kubernetes
     }
 
     @Bean
